+++ test/pylib/tablets.py
@@ -88,3 +88,16 @@ async def get_tablet_info(manager: ManagerClient, server: ServerInfo, keyspace_n
         if row.last_token >= token:
             return row
     return None
+
+async def get_tablet_count(manager: ManagerClient, server: ServerInfo, keyspace_name: str, table_name: str):
+    host = manager.cql.cluster.metadata.get_host(server.ip_addr)
+
+    # read_barrier is needed to ensure that local tablet metadata on the queried node
+    # reflects the finalized tablet movement.
+    await read_barrier(manager.api, server.ip_addr)
+
+    table_id = await manager.get_table_or_view_id(keyspace_name, table_name)
+    table_id = await get_base_table(manager, table_id)
+    rows = await manager.cql.run_async(f"SELECT tablet_count FROM system.tablets where "
+                                       f"table_id = {table_id}", host=host)
+    return rows[0].tablet_count
